Improved finding of importers
 - Now found through a get_importers() function
 - This function is now used in import.py and in the tests - so the tests more accurately reflect the real-worl
 - This function currently returns a list of parsers that are included in pepys-import - later this can be extended to dynamically load parsers
 - Tests have been updated to work with the new get_importers() function

This also found a bug in the way that paths were passed around to importers: just the filename was given to a parser rather than the full path. This was fine
for parsers that can use the file_contents variable to load the file, but the XML parser used in the GPX parser needs to load directly from a file due to
encoding issues, so it actually needs the full path. I've changed file_processor to pass the full_path to the parser.