Remove confusing unused argument to FileProcessor
In import.py we were passing a filename to the FileProcessor when actually it was using the data_store passed to it - so this was just confusing. Now passing nothing,
which means it will use an in-memory database by default, but then actually uses the data_store that's passed to it.