Moved common parts of parser load_this_file method to init_load_this_file method
Rather than having this as a method on the base class and calling with super()
we now have a separate method that is called directly by the file processor
before it calls load_this_file. This means parser writers can't forget
to call the super() method when writing their parser.