Add ability to turn off recording of extractions in an importer
This allows you to leave the `record` calls in place (saves effort of removing them,
and means that you can easily turn them on again in the future if you need to) and have them
basically turn into a no-op. This will also speed up the processing of files that don't need
extractions recorded - as if the record calls are turned off, none of the intermediate representations
will be created either.

Specifically, this involves:

 - A new method on the base Importer class called disable_recording() that can be called to turn off recording for that importer
 (it can be called from __init__ or even within an if statement in _load_this_file to make a decision based on the file name!)
 - A new attribute for HighlightedFile that keeps a list of importers to ignore for that file (so we can ignore just one of the
 three importers that operate on REP files for instance)
 - New logic in the record method for Line and Token objects to tell them to immediately return if the name of the importer in the record call
 matches the name of an importer they're told to ignore