Updated create_shortcuts to delete existing shortcuts first
The shortcuts in both the Send To and Start Menu folders are deleted first (if they exist) and then recreated.

This means that a) we don't get errors when we run create_shortcuts for a second time, and b) we can release a new version of Pepys that has changed some of the shortcuts, and the old ones will be deleted and the new ones created.