Update create_shortcuts to add bin dir to PATH, and set_paths to be able to be called from anywhere
create_shortcuts now adds the bin dir of the Pepys installation to the User's PATH variable (note: the User's PATH variable is different to the system PATH variable - and we can edit it without needing admin rights. Both the User and the System PATH are concatenated to get %PATH% which is actually used as the PATH.

Also, set_paths.bat didn't previously work if it was called outside of the directory that it is in - this now works when called from anywhere.