Add current directory and pip 517 directory to sys.path for deployed version
The embedded version of Python that we're using for deployment works differently to a standard Python installation. Normally the current directory is added to the sys.path variable automatically, but that isn't the case in embedded Python. This causes problems when installation various packages via pip: specifically, distlib (which we got around by including a wheel in the repo) and alembic (which we could get around a similar way, but this is a better way to solve it)

This PR makes some changes to get around that, and get that functionality even in the embedded Python.

1. We add a new .pth file to site-packages which does two things: adds an empty entry to the first element of sys.path and adds a specific folder (pip\_vendor\pip517) to the path too. The first of these fixes things for distlib, and the second allows the alembic install to work correctly (I'd have thought the alembic install would have been fixed by  the first change too, but it isn't - so we had to add the directory manually)

2. We add 'import site' to the main .pth file to tell it to look for other .pth files (we can't do the modifications we did above in the main .pth file as import statements aren't allowed there).

3. We install from pip without using cached files. This takes a bit longer, but means we don't get hidden errors where things work on my computer (because of the cache that I already have from other work) but don't on someone else's.

4. We no longer install the distlib wheel file manually, and delete it from the repository.