Change DefaultResolver to return same Sensor object if resolve_sensor() called multiple times with same args
Previously DefaultResolver would add a new sensor object with the same default sensor name (SENSOR-1) when called
multiple times with the same arguments. This doesn't realistically emulate the behaviour of the CommandLineResolver,
as the user would search for a previously-added sensor and return that instead.

Previously this behaviour was hidden because we always called resolve_sensor() with a sensor_name being given, so it
returned previously stored Sensor objects: this issue only becomes apparent when the sensor name is None, and so it has to be resolved rather than
just found.

Test has been added to ensure that multiple calls with the same arguments will provide the same Sensor object.

This fixes various tests in test_data_store_export_datafile.py, which failed due to multiple sensor entries being added.