Updated CommandLineResolver and relevant tests to pass host_id around when dealing with sensors
In an earlier commit we added host_id as an argument for most sensor-related functions, as we need
to only search/add/show sensors that are related to the current platform. This commit has changes to
ensure that host_id is always passed, and updates the tests so that we have valid platforms to use to get a host_id
before we pass it to the functions under test.