Add get_cached_sensor method to Importer base class, and switch REP and NMEA importers to use it
This is required because these two importers called platform.get_sensor with both sensor_name and sensor_type set to None. That meant that the user would be asked to
specify a sensor _each time a line was parsed_ as there was no way of looking one up. The get_cached_sensor method stores a cache mapping a platform to a sensor, and looks
up in there first, and only resolves if it can't find it.