Change data_store.search_sensor to take platform_id arg, and remove populate_measurement function
The previous `search_sensor` method only took a name to search by - but sensors must always be searched for by platform, as you can have duplicate names for sensors on different platforms. This function has now been changed to take a `platform_id` argument too, and all usages of it changed. These usages were nearly all in the tests - where it was fairly simple to add a search for a platform first, and then use the new search_sensor function to find based on that platform. However, the one place where this wasn't easy to change was in `populate_measurements` which is used to populate some example measurement data (States, Contacts etc). This was available through Pepys Admin, but never used - and its only other uses were in the tests where it wasn't needed - so in discussion with Ian this was removed. Because this was the only function that used the add_to_states method, this was removed too.