Fix command-line resolver to deal properly with a platform name that isn't originally specified
Many importers know a good guess at the platform name from the file, and thus send a platform_name to the resolver, which the user can then
use as a guide to choosing which entry to link it to. However, the NMEA resolver doesn't do this - there is no information on the platform
in the NMEA file, so we call the resolver without specify a platform_name. In this case, the resolver was failing because it was missing all
options provided in the if statements, and thus returning nothing (ie. None).

This change updates the if statements to cover all possibilities, and to separate out bits into nested if statements to make them clearer and
easier to understand.