Added get_cached_platform method to Importer, and change all importers to use it
This cache links the platform name given in the file to a platform in the database. This means that when NELSON is found in a file,
it will need resolving the first time, but can then be looked up from the cache, to save asking the user each time what NELSON is.
This is very similar to what we do for sensors already.