Deal with issue with --training tests where old training config stuck around
This isn't a problem for normal use, as we start a process to run Pepys (with or without
--training), run it, and then finish the process. If we have to change the PEPYS_CONFIG_FILE
env var to deal with --training, then it only takes effect for that process.

However, if we're running in Pytest, we have a single process that keeps going for all the tests,
and so the changed env var (and config resulting from it) sticks around.

These changes reset the PEPYS_CONFIG_FILE env var at the end of the --training tests, and also
change the cli.py files (for Pepys Admin and Pepys Import) to reload the config whenever they are
run, in case the PEPYS_CONFIG_FILE env var has changed.