Move archive_path and local_parsers to be arguments to FileProcessor
These used to be obtained from config.ARCHIVE_PATH and config.LOCAL_PARSERS
inside the FileProcessor. However, when this was the case we needed to
reload the config module inside FileProcessor to pickup possible
changes caused by going into training mode, and therefore getting a different
config file.

However, this meant we couldn't patch config file values inside FileProcessor,
as the `reload(config)` line undid the patch. The only way I could find around
this was to extract these as arguments from FileProcessor and pass them in
with the config file values in cli.py