Fixed bug with filtering by platforms when viewing States in Postgres.
Was caused by the Postgres relationship definition not including the info dict required. I tried to switch to using a relationship just defined in the common_db.py file
but this wasn't possible as this relationship definition requires db-specific bits (ie. a schema for Postgres, but not for SQLite). Therefore, this commit moves the relationships
that have 'secondary' values into the sqlite_db and postgres_db files.