Fix broken merging functionality: make HelpTexts table merge properly, and revert to old way of specifying schema
Note: Originally the Postgres schema was specified on each table definition, but then it was moved to the main Metadata instance.
However, it seems that defining it on the Metadata instance fails in certain situations. All the merging tests passed, but
in actual use with Postgres databases it failed to search for tables in the right schemas when it was merging an object that originally
came from the SQLite database. Adding the schema information back to each table seems to fix this.