Remove unnecessary joined SQL query for Platforms to get the list of participations.
This was being applied to _each_ Platform object every time it was obtained, which slowed down the import process a lot - causing our benchmark tests to fail. It was originally
needed to stop an error in the Maintenance GUI when viewing Platforms, but isn't needed now as we don't show the participations relationship in the edit dialog.