Change how local validators are imported, so that tests that do imports that run after the config tests don't fail.
Any import running as part of a test that ran _after_ the config file tests was failing because of 'left-over' config from the config file tests
that had set up validators that deliberately failed. This changes how the validators are imported so that they pick up the new config after the config
tests have run. This _could_ cause a slowdown in some situations, but I don't believe local validators are even used by the client yet, so I don't think
it'll be a problem.