Change benchmark so that FileProcessor is initialised once for each test, rather than an instance being shared.
This slows the test down a bit (though means that more of the FileProcessor is being benchmarked), but stops us adding extra directories to the FileProcessor output path each time it is called.
This was happening before (even on Python 3.8), and meant that you had a massive long path like D:\\a\\pepys-import\\pepys-import\\tests\\benchmarks\\benchmark_data\\output\\2021\\11\\13\\15\\53\\13\\2021\\11\\13\\15\\53\\13\\2021\\11\\13\\15\\53\\14\\2021\\11\\13\\15\\53\\14\\2021\\11\\13\\15\\53\\15\\2021\\11\\13\\15\\53\\15\\2021\\11\\13\\15\\53\\16\\2021\\11\\13\\15\\53\\16\\reports\\rep_test1_highlighted.html

My guess is that Python 3.9's speed improvements have made the benchmarking do more iterations (to get an accurate time), and this has caused the path to become too many levels for Windows to cope
with.